% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risk.R
\name{printRisk}
\alias{printRisk}
\title{Print calculated ccalues of VaR or ES}
\usage{
printRisk(data, time, typ)
}
\arguments{
\item{data}{A dataframe object with column dailyReturnRate}

\item{time}{A number of time window.}

\item{typ}{A character - VaR or ES}
}
\value{
calculated values of VaR or ES
}
\description{
Print calculated ccalues of VaR or ES
}
\examples{
#data("returnRateEUR")
#printRisk(returnRateEUR, 500, typ="VaR")
}
\keyword{Risk}
