% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risk.R
\name{Risk}
\alias{Risk}
\title{Calculating for given rates of return the value at risk and the expected loss using the historical method.}
\usage{
Risk(data, time)
}
\arguments{
\item{data}{A dataframe object with column dailyReturnRate}

\item{time}{A number of time window.}
}
\value{
A data frame with calculated VaR and ES
}
\description{
Calculating for given rates of return the value at risk and the expected loss using the historical method.
}
\examples{
#data("returnRateEUR")
#Risk(returnRateEUR, 200)
}
\keyword{Risk}
